web/requirements.txt
flask==3.0.2
Flask-SSE==1.0.0
gunicorn==21.2.0
gevent==23.9.1
langchain==0.1.12
langchain-core==0.1.32
langchain-nvidia-ai-endpoints==0.0.16
langchain-openai==0.0.8
pandas==2.0.3
redis==5.0.1
python-dotenv
sentencepiece==0.1.97
